Update cs_CZ.json
translated lines 1068 - 1111

I'd like to inform you that there are 3 phases of plural in Czech. Your json file is not fully capable to cover it. This is general issue with localization (I know it from my indie games). Not only czech, but all slovan languages have it. Arabic as well. 

Example:
In english there are seconds (0, 2-infinity) and a second (1).
In czech there is 1 second [vterina], 2,3,4 seconds [vteriny] and 5-infinity (different term) seconds [vterin].

So for now I translated it in a way that only number 1 will sounds weird. 

Also czech is more gender based language. 
Example: Stay signed it in English is covering both men and women. 
In czech we have Stay signed in written differently for men and women. 
Zustat prihlasen [telling to a man] vs. Zustat prihlasena [telling to a woman]. For now I used universal way which sound more artistical/old-school (understand bookish).