Update cs_CZ.json
Translated lines 2000 - 2102. I have several questions: Chris what is Card ending on line 2051? Is this the code on the back of each card?
Also, what does "at" on line 2096 mean? I believe it is connected to the new exchange rate and connected with the previous 2 lines. I believe there is no reposition in CZ.