Update nl_NL.json
1) Second audit till line 918
2) I was also looking at the reason why the lines of the Dutch translation did not match up exactly with the English text: https://github.com/nash-io/locale/blob/master/locales/en.json    
Line 957 of the English text is entirely empty and this thus shifts all lines further down to differ 1 from the Dutch text (and the same is probably true for other translations). A correction in the English text of this thus seems desirable to prevent the chance for errors in other translations.